refactor: move role and authentication to security
