+++ src/main/java/tw/commonground/backend/security/UserRole.java
@@ -1,0 +1,5 @@
+package tw.commonground.backend.security;
+
+public enum UserRole {
+    ROLE_ADMIN, ROLE_USER, ROLE_NOT_SETUP, ROLE_SERVICE_ACCOUNT_READ
+}
