+++ src/main/java/tw/commonground/backend/service/user/entity/FullUserEntity.java
@@ -1,5 +1,7 @@
 package tw.commonground.backend.service.user.entity;
 
+import tw.commonground.backend.security.UserRole;
+
 import java.util.UUID;
 
 public interface FullUserEntity {
