+++ src/main/java/tw/commonground/backend/service/user/entity/UserRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
+import tw.commonground.backend.security.UserRole;
 
 import java.util.List;
 import java.util.Optional;
