--- src/main/java/tw/commonground/backend/service/user/entity/UserRole.java
@@ -1,5 +1,0 @@
-package tw.commonground.backend.service.user.entity;
-
-public enum UserRole {
-    ROLE_ADMIN, ROLE_USER, ROLE_NOT_SETUP
-}
